+++ xchange-acx/src/main/java/org/knowm/xchange/acx/dto/account/AcxAccount.java
@@ -1,19 +1,18 @@
 package org.knowm.xchange.acx.dto.account;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.math.BigDecimal;
 
 public class AcxAccount {
   public final String currency;
   /** Account balance, exclude locked funds */
-  public final BigDecimal balance;
+  public final Double balance;
   /** locked funds */
-  public final BigDecimal locked;
+  public final Double locked;
 
   public AcxAccount(
       @JsonProperty("currency") String currency,
-      @JsonProperty("balance") BigDecimal balance,
-      @JsonProperty("locked") BigDecimal locked) {
+      @JsonProperty("balance") Double balance,
+      @JsonProperty("locked") Double locked) {
     this.currency = currency;
     this.balance = balance;
     this.locked = locked;
