+++ xchange-acx/src/main/java/org/knowm/xchange/acx/dto/marketdata/AcxTicker.java
@@ -1,32 +1,31 @@
 package org.knowm.xchange.acx.dto.marketdata;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.math.BigDecimal;
 
 public class AcxTicker {
   /** Current sell price */
-  public final BigDecimal buy;
+  public final Double buy;
   /** Current sell price */
-  public final BigDecimal sell;
+  public final Double sell;
 
-  public final BigDecimal open;
+  public final Double open;
   /** Lowest price in last 24 hours */
-  public final BigDecimal low;
+  public final Double low;
   /** Highest price in last 24 hours */
-  public final BigDecimal high;
+  public final Double high;
   /** Last price */
-  public final BigDecimal last;
+  public final Double last;
   /** Trade volume in last 24 hours */
-  public final BigDecimal vol;
+  public final Double vol;
 
   public AcxTicker(
-      @JsonProperty("buy") BigDecimal buy,
-      @JsonProperty("sell") BigDecimal sell,
-      @JsonProperty("open") BigDecimal open,
-      @JsonProperty("low") BigDecimal low,
-      @JsonProperty("high") BigDecimal high,
-      @JsonProperty("last") BigDecimal last,
-      @JsonProperty("vol") BigDecimal vol) {
+      @JsonProperty("buy") Double buy,
+      @JsonProperty("sell") Double sell,
+      @JsonProperty("open") Double open,
+      @JsonProperty("low") Double low,
+      @JsonProperty("high") Double high,
+      @JsonProperty("last") Double last,
+      @JsonProperty("vol") Double vol) {
     this.buy = buy;
     this.sell = sell;
     this.open = open;
