+++ xchange-anx/src/test/java/org/knowm/xchange/anx/v2/dto/account/WalletJSONTest.java
@@ -7,7 +7,6 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.io.InputStream;
-import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
@@ -38,22 +37,22 @@ public class WalletJSONTest {
 
     // Get Balance
     assertThat(anxAccountInfo.getWallets().get("BTC").getBalance().getValue())
-        .isEqualTo(new BigDecimal("100000.01988000"));
+        .isEqualTo(new Double("100000.01988000"));
     assertThat(anxAccountInfo.getWallets().get("USD").getBalance().getValue())
-        .isEqualTo(new BigDecimal("100000.00000"));
+        .isEqualTo(new Double("100000.00000"));
     assertThat(anxAccountInfo.getWallets().get("HKD").getBalance().getValue())
-        .isEqualTo(new BigDecimal("99863.07000"));
+        .isEqualTo(new Double("99863.07000"));
 
     assertThat(anxAccountInfo.getWallets().get("LTC").getBalance().getValue())
-        .isEqualTo(new BigDecimal("100000.00000000"));
+        .isEqualTo(new Double("100000.00000000"));
     assertThat(anxAccountInfo.getWallets().get("DOGE").getBalance().getValue())
-        .isEqualTo(new BigDecimal("9999781.09457936"));
+        .isEqualTo(new Double("9999781.09457936"));
 
     // Get Other Balance
     assertThat(anxAccountInfo.getWallets().get("BTC").getMaxWithdraw().getValue())
-        .isEqualTo(new BigDecimal("20.00000000"));
+        .isEqualTo(new Double("20.00000000"));
     assertThat(anxAccountInfo.getWallets().get("BTC").getDailyWithdrawLimit().getValue())
-        .isEqualTo(new BigDecimal("20.00000000"));
+        .isEqualTo(new Double("20.00000000"));
   }
 
   @Test
